// src/quest/interface.ts
export interface AllQuestsResponse {
	/**
	 * The current quests for the user
	 */
	quests: Quest[];
	/**
	 * The quests that the user cannot participate in
	 */
	excluded_quests: Partial<Quest>[];
	/**
	 * When the user can enroll in quests again
	 */
	quest_enrollment_blocked_until: string | null;
}

export type Snowflake = string;

export interface Quest {
	/**
	 * The ID of the quest
	 */
	id: Snowflake;
	/**
	 * The configuration and metadata for the quest
	 */
	config: QuestConfig;
	/**
	 * The user's quest progress, if it has been accepted
	 */
	user_status: QuestUserStatus | null;
	/**
	 * The content areas where the quest can be shown
	 * @deprecated
	 */
	targeted_content: number;
	/**
	 * Whether the quest is unreleased and in preview for Discord employees
	 */
	preview: boolean;
	// ?
	traffic_metadata_raw?: string;
	traffic_metadata_sealed?: string;
}

/**
 * The config structure has multiple distinct versions with different field sets. Only actively used versions are kept documented. As of now, only the latest version is available.
 */
export interface QuestConfig {
	/**
	 * The ID of the quest
	 */
	id: Snowflake;
	/**
	 * Quest configuration version
	 */
	config_version: number;
	/**
	 * When the quest period starts
	 */
	starts_at: string;
	/**
	 * When the quest period ends
	 */
	expires_at: string;
	/**
	 * The quest features enabled for the quest
	 */
	features: number;
	/**
	 * The application metadata for the quest
	 */
	application: QuestApplication;
	/**
	 * Object that holds the quest's assets
	 */
	assets: QuestAssets;
	/**
	 * The accent colors for the quest
	 */
	colors: QuestGradient;
	/**
	 * Human-readable metadata for the quest
	 */
	messages: QuestMessages;
	/**
	 * The task configuration for the quest
	 */
	// task_config: QuestTaskConfig;
	/**
	 * Specifies rewards for the quest (e.g. collectibles)
	 */
	rewards_config: QuestRewardsConfig;
	/**
	 * The configuration for the video quest
	 */
	video_metadata?: QuestVideoMetadata;
	/**
	 * The configuration for the quest co-sponsor
	 */
	cosponsor_metadata?: QuestCosponsorMetadata;
	// ?
	task_config_v2: {
		tasks: /*{
			WATCH_VIDEO?: {
				type: QuestTaskConfigType.WATCH_VIDEO;
				target: number;
			};
			WATCH_VIDEO_ON_MOBILE?: {
				type: QuestTaskConfigType.WATCH_VIDEO_ON_MOBILE;
				target: number;
			};
			PLAY_ON_DESKTOP?: {
				type: QuestTaskConfigType.PLAY_ON_DESKTOP;
				target: number;
			};
			PLAY_ON_XBOX?: {
				type: QuestTaskConfigType.PLAY_ON_XBOX;
				target: number;
			};
			PLAY_ON_PLAYSTATION?: {
				type: QuestTaskConfigType.PLAY_ON_PLAYSTATION;
				target: number;
			};
			PLAY_ACTIVITY?: {
				type: QuestTaskConfigType.PLAY_ACTIVITY;
				target: number;
			};
			ACHIEVEMENT_IN_ACTIVITY?: {
				type: QuestTaskConfigType.ACHIEVEMENT_IN_ACTIVITY;
				target: number;
				event_name: 'progress';
				applications: {
					id: string;
				}[];
			};
		};
		*/
		Record<
			QuestTaskConfigType,
			{ type: QuestTaskConfigType; target: number }
		>;
	};
}

export interface QuestUserStatus {
	/**
	 * The ID of the user
	 */
	user_id: Snowflake;
	/**
	 * The ID of the quest
	 */
	quest_id?: Snowflake;
	/**
	 * When the user accepted the quest
	 */
	enrolled_at: string | null;
	/**
	 * When the user completed the quest
	 */
	completed_at: string | null;
	/**
	 * When the user claimed the quest's reward
	 */
	claimed_at: string | null;
	/**
	 * Which reward tier the user has claimed, if the quest's assignment_method is TIERED
	 */
	claimed_tier?: number | null;
	/**
	 * When the last heartbeat was received
	 */
	last_stream_heartbeat_at?: string | null;
	/**
	 * Duration (in seconds) the user has streamed the game for since the quest was accepted
	 */
	stream_progress_seconds?: string;
	/**
	 * The content areas the user has dismissed for the quest
	 */
	dismissed_quest_content?: number;
	/**
	 * The user's progress for each task in the quest, keyed by their event name
	 */
	progress: Record<string, QuestTaskProgress>;
}
// ...
export enum QuestTaskConfigType {
	STREAM_ON_DESKTOP = 'STREAM_ON_DESKTOP',
	PLAY_ON_DESKTOP = 'PLAY_ON_DESKTOP',
	PLAY_ON_DESKTOP_V2 = 'PLAY_ON_DESKTOP_V2',
	PLAY_ON_XBOX = 'PLAY_ON_XBOX',
	PLAY_ON_PLAYSTATION = 'PLAY_ON_PLAYSTATION',
	WATCH_VIDEO = 'WATCH_VIDEO',
	WATCH_VIDEO_ON_MOBILE = 'WATCH_VIDEO_ON_MOBILE',
	PLAY_ACTIVITY = 'PLAY_ACTIVITY',
	ACHIEVEMENT_IN_GAME = 'ACHIEVEMENT_IN_GAME',
	ACHIEVEMENT_IN_ACTIVITY = 'ACHIEVEMENT_IN_ACTIVITY',
}